@@ -30,17 +30,17 @@
     "@mui/material": "5.11.12",
     "@mui/x-date-pickers": "6.19.6",
     "@radix-ui/react-dialog": "^1.1.7",
-    "@supabase/supabase-js": "^2.49.1",
+    "@supabase/supabase-js": "2.50.2",
     "@tabler/icons-react": "^3.31.0",
     "@tanstack/react-query": "^5.17.19",
     "@types/bcrypt": "^5.0.2",
     "@types/cors": "^2.8.17",
     "@types/dotenv": "^8.2.3",
-    "@types/express": "^4.17.21",
     "@types/multer": "^1.4.11",
     "@types/pg": "^8.11.11",
     "@types/uuid": "^10.0.0",
     "ai": "4.3.16",
+    "axios": "1.10.0",
     "bcrypt": "^5.1.1",
     "chart.js": "^3.7.1",
     "class-variance-authority": "^0.7.1",
@@ -51,7 +51,7 @@
     "date-fns": "^2.30.0",
     "dotenv": "^16.4.7",
     "exceljs": "^4.4.0",
-    "express": "^4.18.3",
+    "express": "4.21.2",
     "express-rate-limit": "^7.5.0",
     "file-saver": "^2.0.5",
     "formidable": "^3.5.4",
@@ -89,8 +89,10 @@
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.3.0",
     "@testing-library/user-event": "^14.6.1",
+    "@types/express": "5.0.3",
     "@types/file-saver": "^2.0.7",
     "@types/formidable": "^3.4.5",
+    "@types/node": "24.0.4",
     "@types/react": "^18.2.43",
     "@types/react-csv": "^1.1.10",
     "@types/react-dom": "^18.2.17",
@@ -104,9 +106,11 @@
     "eslint-plugin-react-refresh": "^0.4.5",
     "jsdom": "^26.1.0",
     "postcss": "^8.5.3",
+    "supabase": "2.26.9",
     "tailwindcss": "^3.4.4",
     "terser": "^5.41.0",
-    "typescript": "^5.2.2",
+    "ts-node": "10.9.2",
+    "typescript": "5.8.3",
     "vite": "^5.0.8",
     "vitest": "^3.1.4"
   }
