@@ -1,4 +1,8 @@
 import axios from 'axios';
+import { format } from 'date-fns';
+import { AppointmentData, WhatsAppResponse } from '../../shared/types';
+
+const WHATSAPP_API_VERSION = 'v17.0';
 
 interface WhatsAppMessage {
   to: string;
@@ -7,58 +11,57 @@ interface WhatsAppMessage {
     language: {
       code: string;
     };
-    components: any[];
+    components: {
+      type: string;
+      parameters: {
+        type: string;
+        text: string;
+      }[];
+    }[];
   };
 }
 
-const WHATSAPP_API_VERSION = 'v17.0';
-const PHONE_NUMBER_ID = process.env.WHATSAPP_PHONE_NUMBER_ID;
-const ACCESS_TOKEN = process.env.WHATSAPP_ACCESS_TOKEN;
-
 export class WhatsAppService {
-  private static formatPhoneNumber(phoneNumber: string): string {
-    // Remove any non-digit characters
-    const cleaned = phoneNumber.replace(/\D/g, '');
-    
-    // If number doesn't start with country code (91), add it
-    if (!cleaned.startsWith('91')) {
-      return `91${cleaned}`;
+  private static formatPhoneNumber(phone: string): string {
+    let formatted = phone.replace(/\D/g, '');
+    if (!formatted.startsWith('91') && formatted.length === 10) {
+      formatted = '91' + formatted;
     }
-    
-    return cleaned;
+    return formatted;
+  }
+
+  private static formatDateTime(date: Date): string {
+    return format(date, 'PPpp'); // e.g., "Apr 29, 2023, 3:00 PM"
+  }
+
+  private static formatDate(date: Date): string {
+    return format(date, 'PPP'); // e.g., "April 29th, 2023"
   }
 
   private static formatCurrency(amount: number): string {
     return new Intl.NumberFormat('en-IN', {
       style: 'currency',
-      currency: 'INR',
-      minimumFractionDigits: 0,
-      maximumFractionDigits: 0
+      currency: 'INR'
     }).format(amount);
   }
 
-  private static formatDate(date: Date): string {
-    return date.toLocaleDateString('en-IN', {
-      weekday: 'long',
-      year: 'numeric',
-      month: 'long',
-      day: 'numeric'
-    });
+  private static formatDuration(minutes: number): string {
+    const hours = Math.floor(minutes / 60);
+    const remainingMinutes = minutes % 60;
+    
+    if (hours > 0) {
+      return `${hours} hr${hours > 1 ? 's' : ''} ${remainingMinutes > 0 ? `${remainingMinutes} min` : ''}`;
+    }
+    return `${remainingMinutes} min`;
   }
 
-  private static formatTime(date: Date): string {
-    return date.toLocaleTimeString('en-IN', {
-      hour: '2-digit',
-      minute: '2-digit',
-      hour12: true
-    }).toUpperCase();
+  private static formatServicesList(services: Array<{ name: string; price: number; duration: number }>): string {
+    return services.map(service => 
+      `- ${service.name} (${this.formatDuration(service.duration)}) - ${this.formatCurrency(service.price)}`
+    ).join('\n');
   }
 
-  private static formatDateTime(date: Date): string {
-    return `${this.formatDate(date)} at ${this.formatTime(date)}`;
-  }
-
-  private static async sendMessage(message: WhatsAppMessage): Promise<any> {
+  private static async sendMessage(message: WhatsAppMessage): Promise<WhatsAppResponse> {
     try {
       // Format the phone number before sending
       message.to = this.formatPhoneNumber(message.to);
@@ -84,51 +87,61 @@ export class WhatsAppService {
         }
       );
 
-      return response.data;
+      return {
+        success: true,
+        data: response.data
+      };
     } catch (error) {
       if (axios.isAxiosError(error)) {
         const errorCode = error.response?.data?.error?.code;
         const errorMessage = error.response?.data?.error?.message;
         
         // Handle specific WhatsApp API errors
+        const errorResponse = {
+          success: false,
+          error: {
+            code: errorCode,
+            message: errorMessage
+          }
+        };
+
         switch (errorCode) {
           case 132000:
-            throw new Error(`WhatsApp opt-in required: The recipient (${message.to}) needs to message the business first. Please ask the client to send a message to our WhatsApp business number.`);
+            errorResponse.error.message = `WhatsApp opt-in required: The recipient (${message.to}) needs to message the business first.`;
+            break;
           case 131047:
-            throw new Error(`Template error: ${errorMessage}. Please check if the template parameters are correct.`);
+            errorResponse.error.message = `Template error: ${errorMessage}. Please check template parameters.`;
+            break;
           case 130429:
-            throw new Error('Rate limit exceeded. Please try again later.');
+            errorResponse.error.message = 'Rate limit exceeded. Please try again later.';
+            break;
           case 100:
-            throw new Error(`Invalid parameter: ${errorMessage}. Please check the phone number format and other parameters.`);
-          default:
-            throw new Error(`WhatsApp API error (${errorCode}): ${errorMessage}`);
+            errorResponse.error.message = `Invalid parameter: ${errorMessage}. Please check phone number format.`;
+            break;
         }
+        
+        return errorResponse;
       }
-      throw error;
+      return {
+        success: false,
+        error: {
+          message: error instanceof Error ? error.message : 'Unknown error occurred'
+        }
+      };
     }
   }
 
-  static async sendAppointmentConfirmation(
-    phoneNumber: string,
-    clientName: string,
-    appointmentDate: Date,
-    services: { name: string; price: number; duration: number }[],
-    stylist: string
-  ) {
-    const message: WhatsAppMessage = {
+  static async sendTestMessage(phoneNumber: string): Promise<WhatsAppResponse> {
+    const message = {
       to: phoneNumber,
       template: {
-        name: 'appointment_confirm',
+        name: 'test_message',
         language: { code: 'en' },
         components: [
           {
             type: 'body',
             parameters: [
-              { type: 'text', text: clientName },
-              { type: 'text', text: this.formatDateTime(appointmentDate) },
-              { type: 'text', text: services[0].name },
-              { type: 'text', text: stylist },
-              { type: 'text', text: this.formatCurrency(services[0].price) }
+              { type: 'text', text: '🎉 Hello! This is a test message from RG Salon\'s WhatsApp integration. If you receive this, our messaging system is working correctly! 💈' }
             ]
           }
         ]
@@ -138,24 +151,22 @@ export class WhatsAppService {
     return this.sendMessage(message);
   }
 
-  static async sendSpecialOffer(
-    phoneNumber: string,
-    clientName: string,
-    offerDetails: string,
-    validUntil: Date
-  ) {
+  static async sendAppointmentConfirmation(appointment: AppointmentData): Promise<WhatsAppResponse> {
     const message: WhatsAppMessage = {
-      to: phoneNumber,
+      to: appointment.client.phone,
       template: {
-        name: 'rg_salon_special_offer',
+        name: 'appointment_confirmation',
         language: { code: 'en' },
         components: [
           {
             type: 'body',
             parameters: [
-              { type: 'text', text: clientName },
-              { type: 'text', text: offerDetails },
-              { type: 'text', text: this.formatDate(validUntil) }
+              { type: 'text', text: appointment.client.name },
+              { type: 'text', text: format(appointment.datetime, 'PPpp') },
+              { type: 'text', text: appointment.stylist.name },
+              { type: 'text', text: appointment.services.map(s => s.name).join(', ') },
+              { type: 'text', text: format(appointment.services.reduce((total, s) => total + s.duration, 0), 'H:mm') },
+              { type: 'text', text: appointment.services.reduce((total, s) => total + s.price, 0).toFixed(2) }
             ]
           }
         ]
@@ -165,32 +176,24 @@ export class WhatsAppService {
     return this.sendMessage(message);
   }
 
-  static async sendAppointmentUpdate(
-    phoneNumber: string,
-    clientName: string,
-    oldDate: Date,
-    newDate: Date,
-    services: { name: string; price: number; duration: number }[],
-    stylist: string,
-    totalAmount: number
-  ) {
-    const serviceText = `${services[0].name} (${services[0].duration} min) - ${this.formatCurrency(services[0].price)}`;
-
+  static async sendAppointmentReschedule(
+    appointment: AppointmentData,
+    oldDateTime: Date
+  ): Promise<WhatsAppResponse> {
     const message: WhatsAppMessage = {
-      to: phoneNumber,
+      to: appointment.client.phone,
       template: {
-        name: 'appointment_updation',
+        name: 'appointment_reschedule',
         language: { code: 'en' },
         components: [
           {
             type: 'body',
             parameters: [
-              { type: 'text', text: clientName },
-              { type: 'text', text: `${this.formatDate(oldDate)} at ${this.formatTime(oldDate)}` },
-              { type: 'text', text: `${this.formatDate(newDate)} at ${this.formatTime(newDate)}` },
-              { type: 'text', text: serviceText },
-              { type: 'text', text: stylist },
-              { type: 'text', text: this.formatCurrency(totalAmount) }
+              { type: 'text', text: appointment.client.name },
+              { type: 'text', text: format(oldDateTime, 'PPpp') },
+              { type: 'text', text: format(appointment.datetime, 'PPpp') },
+              { type: 'text', text: appointment.stylist.name },
+              { type: 'text', text: appointment.services.map(s => s.name).join(', ') }
             ]
           }
         ]
@@ -200,14 +203,9 @@ export class WhatsAppService {
     return this.sendMessage(message);
   }
 
-  static async sendAppointmentReminder(
-    phoneNumber: string,
-    clientName: string,
-    appointmentDate: Date,
-    services: { name: string; price: number; duration: number }[]
-  ) {
+  static async sendAppointmentReminder(appointment: AppointmentData): Promise<WhatsAppResponse> {
     const message: WhatsAppMessage = {
-      to: phoneNumber,
+      to: appointment.client.phone,
       template: {
         name: 'appointment_reminder',
         language: { code: 'en' },
@@ -215,9 +213,9 @@ export class WhatsAppService {
           {
             type: 'body',
             parameters: [
-              { type: 'text', text: clientName },
-              { type: 'text', text: this.formatDateTime(appointmentDate) },
-              { type: 'text', text: services[0].name }
+              { type: 'text', text: appointment.client.name },
+              { type: 'text', text: format(appointment.datetime, 'PPpp') },
+              { type: 'text', text: appointment.services.map(s => s.name).join(', ') }
             ]
           }
         ]
@@ -226,33 +224,4 @@ export class WhatsAppService {
 
     return this.sendMessage(message);
   }
-
-  static async sendAppointmentCancellation(
-    phoneNumber: string,
-    clientName: string,
-    appointmentDate: Date
-  ) {
-    // For cancellation, we'll use the special offer template with cancellation message
-    const message: WhatsAppMessage = {
-      to: phoneNumber,
-      template: {
-        name: 'rg_salon_special_offer',
-        language: { code: 'en' },
-        components: [
-          {
-            type: 'body',
-            parameters: [
-              { type: 'text', text: clientName },
-              { type: 'text', text: `Your appointment scheduled for ${this.formatDateTime(appointmentDate)} has been cancelled. Please contact us to reschedule.` },
-              { type: 'text', text: this.formatDate(new Date(Date.now() + 30 * 24 * 60 * 60 * 1000)) } // 30 days from now
-            ]
-          }
-        ]
-      }
-    };
-
-    return this.sendMessage(message);
-  }
-}
-
-export { WhatsAppService }; 
+} 
